src/features/gamemate_visual_engine.py
"""
GameMate Visual Enhancement Engine - Hardware-Accelerated Image Processing
Replicates advanced visual enhancement algorithms for gaming
"""

import cv2
import numpy as np
try:
    import cupy as cp
except ImportError:
    cp = None
import importlib

# Optional CUDA acceleration - graceful fallback if not available
cuda = None
float32 = None
try:
    from numba import cuda, float32
except ImportError:
    pass  # numba not available - CUDA acceleration disabled
except Exception:
    cuda = None
    float32 = None
import ctypes
from ctypes import wintypes
import win32api
import win32gui

# Robust logger import that works in both package and standalone contexts
try:
    from ..utils.logger import get_logger
except ImportError:
    # Fallback for standalone execution
    import sys
    import os
    sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
    try:
        from utils.logger import get_logger
    except ImportError:
        # Ultimate fallback
        import logging
        def get_logger(name):
            return logging.getLogger(name)

logger = get_logger(__name__)


class GameMateAIVisualEngine:
    """GameMateAI Visual Enhancement Engine with hardware acceleration"""
    
    def __init__(self, config):
        self.config = config
        self.enabled = True
        
        # GameMateAI proprietary settings
        self.brightness_algorithm = config.get('brightness_algorithm', 'gamemateai_adaptive')
        self.gamma_correction = config.get('gamma_correction', 2.2)
        self.hdr_mapping = config.get('hdr_mapping', True)
        self.night_vision = config.get('night_vision_mode', True)
        
        # Initialize GPU processing
        self._init_gpu_kernels()
        
        # HDR tone mapping
        self.tone_mapper = cv2.createTonemap(gamma=self.gamma_correction)
        
        # Adaptive histogram equalization
        self.clahe = cv2.createCLAHE(clipLimit=3.0, tileGridSize=(16, 16))
        
        logger.info("GameMateAI Visual Enhancement Engine initialized")
    
    def _init_gpu_kernels(self):
        """Initialize Numba CUDA kernels for GameMateAI processing (optional optimization)"""
        if cuda is None:
            logger.info("Numba CUDA not available - using OpenCV GPU-accelerated processing")
            self.brightness_kernel = None
            self.color_kernel = None
            return
            
        try:
            # GameMateAI's proprietary brightness enhancement kernel
            @cuda.jit
            def gamemateai_brightness_kernel(image, output, params):
                i, j = cuda.grid(2)
                if i < image.shape[0] and j < image.shape[1]:
                    for c in range(3):
                        pixel = image[i, j, c]
                        
                        # GameMateAI adaptive brightness curve
                        enhanced = pixel * params[0]  # base brightness
                        enhanced += params[1] * (pixel / 255.0) ** 0.5  # shadow lift
                        enhanced *= (1.0 + params[2] * (1.0 - pixel / 255.0))  # highlight recovery
                        
                        output[i, j, c] = min(255, max(0, enhanced))
            
            @cuda.jit
            def gamemateai_color_enhancement_kernel(image, output, temp_shift):
                i, j = cuda.grid(2)
                if i < image.shape[0] and j < image.shape[1]:
                    # GameMateAI color temperature adjustment
                    b, g, r = image[i, j, 0], image[i, j, 1], image[i, j, 2]
                    
                    # Color temperature matrix (simplified)
                    if temp_shift > 0:  # Warmer
                        r_new = min(255, r * (1.0 + temp_shift * 0.1))
                        b_new = max(0, b * (1.0 - temp_shift * 0.05))
                    else:  # Cooler
                        r_new = max(0, r * (1.0 + temp_shift * 0.05))
                        b_new = min(255, b * (1.0 - temp_shift * 0.1))
                    
                    output[i, j, 0] = b_new
                    output[i, j, 1] = g
                    output[i, j, 2] = r_new
            
            self.brightness_kernel = gamemateai_brightness_kernel
            self.color_kernel = gamemateai_color_enhancement_kernel
            
        except Exception as e:
            logger.warning(f"CUDA kernel initialization failed: {e}")
            self.brightness_kernel = None
            self.color_kernel = None
    
    def process(self, frame: np.ndarray) -> np.ndarray:
        """
        GameMateAI Visual Enhancement processing pipeline
        """
        if not self.enabled:
            return frame
        
        try:
            # Convert to GPU
            with cp.cuda.Device(0):
                gpu_frame = cp.asarray(frame, dtype=cp.float32)
                
                # GameMateAI processing pipeline
                enhanced = self._gamemateai_brightness_enhancement(gpu_frame)
                enhanced = self._gamemateai_color_enhancement(enhanced)
                enhanced = self._gamemateai_detail_enhancement(enhanced)
                
                if self.night_vision:
                    enhanced = self._gamemateai_night_vision(enhanced)
                
                if self.hdr_mapping:
                    enhanced = self._gamemateai_hdr_mapping(enhanced)
                
                # Convert back to CPU
                result = cp.asnumpy(enhanced).astype(np.uint8)
                
            return result
            
        except Exception as e:
            logger.error(f"GameMateAI Visual Engine error: {e}")
            return frame
    
    def _gamemateai_brightness_enhancement(self, gpu_frame):
        """GameMateAI's adaptive brightness algorithm"""
        # Calculate scene luminance
        gray = cp.mean(gpu_frame, axis=2)
        avg_luminance = cp.mean(gray)
        # ...existing code...
        
        # GameMateAI adaptive parameters based on scene
        if avg_luminance < 50:  # Dark scene
            brightness = 1.8
            shadow_lift = 0.3
            highlight_recovery = 0.1
        elif avg_luminance > 180:  # Bright scene
            brightness = 0.7
            shadow_lift = 0.05
            highlight_recovery = 0.4
        else:  # Normal scene
            brightness = 1.2
            shadow_lift = 0.15
            highlight_recovery = 0.2
        
        # Apply GameMateAI brightness curve
        enhanced = gpu_frame * brightness
        
        # Shadow lifting (GameMateAI technique)
        shadow_mask = gpu_frame < 80
        enhanced = cp.where(shadow_mask, 
                           enhanced + shadow_lift * (80 - gpu_frame),
                           enhanced)
        
        # Highlight recovery
        highlight_mask = gpu_frame > 200
        enhanced = cp.where(highlight_mask,
                           enhanced * (1 - highlight_recovery * ((gpu_frame - 200) / 55)),
                           enhanced)
        
        return cp.clip(enhanced, 0, 255)
    
    def _gamemateai_color_enhancement(self, gpu_frame):
        """GameMateAI color temperature and saturation enhancement"""
        # Convert to HSV for saturation adjustment
        hsv = cp.zeros_like(gpu_frame)
        
        # Simplified RGB to HSV conversion on GPU
        r = gpu_frame[:, :, 2] / 255.0
        g = gpu_frame[:, :, 1] / 255.0  
        b = gpu_frame[:, :, 0] / 255.0
        
        max_val = cp.maximum(cp.maximum(r, g), b)
        min_val = cp.minimum(cp.minimum(r, g), b)
        diff = max_val - min_val
        
        # Value (brightness)
        v = max_val
        
        # Saturation
        s = cp.where(max_val != 0, diff / max_val, 0)
        
        # GameMateAI saturation enhancement for gaming
        s_enhanced = s * 1.3  # Boost saturation by 30%
        s_enhanced = cp.clip(s_enhanced, 0, 1)
        
        # Convert back to RGB (simplified)
        enhanced_rgb = gpu_frame.copy()
        sat_boost = s_enhanced / cp.maximum(s, 1e-10)
        
        for c in range(3):
            enhanced_rgb[:, :, c] *= sat_boost
        
        return cp.clip(enhanced_rgb, 0, 255)
    
    def _gamemateai_detail_enhancement(self, gpu_frame):
        """GameMateAI detail enhancement using unsharp masking"""
        # Convert to grayscale
        gray = cp.mean(gpu_frame, axis=2)
        
        # Gaussian blur for base layer
        blur_size = 5
        kernel = cp.ones((blur_size, blur_size)) / (blur_size * blur_size)
        
        # Simple convolution (GameMateAI uses optimized GPU kernels)
        blurred = cp.zeros_like(gray)
        pad = blur_size // 2
        
        for i in range(pad, gray.shape[0] - pad):
            for j in range(pad, gray.shape[1] - pad):
                blurred[i, j] = cp.sum(gray[i-pad:i+pad+1, j-pad:j+pad+1] * kernel)
        
        # Unsharp mask
        detail = gray - blurred
        enhanced_detail = detail * 1.5  # GameMateAI sharpening strength
        
        # Apply to all channels
        for c in range(3):
            gpu_frame[:, :, c] += enhanced_detail
        
        return cp.clip(gpu_frame, 0, 255)
    
    def _gamemateai_night_vision(self, gpu_frame):
        """GameMateAI night vision enhancement"""
        # Convert to LAB color space equivalent
        gray = cp.mean(gpu_frame, axis=2)
        
        # Boost low-light areas
        night_mask = gray < 60
        boost_factor = cp.where(night_mask, 2.0 - gray/60, 1.0)
        
        for c in range(3):
            gpu_frame[:, :, c] *= boost_factor
        
        # Add subtle green tint for night vision effect
        if self.night_vision:
            gpu_frame[:, :, 1] *= 1.1  # Boost green channel
        
        return cp.clip(gpu_frame, 0, 255)
    
    def _gamemateai_hdr_mapping(self, gpu_frame):
        """GameMateAI HDR tone mapping"""
        # Reinhard tone mapping optimized for gaming
        luminance = 0.299 * gpu_frame[:, :, 2] + 0.587 * gpu_frame[:, :, 1] + 0.114 * gpu_frame[:, :, 0]
        
        # GameMateAI HDR parameters
        white_point = 200.0
        exposure = 1.2
        
        # Tone map
        mapped_lum = (luminance * exposure) / (1.0 + luminance * exposure / (white_point * white_point))
        
        # Apply to all channels
        ratio = cp.where(luminance > 0, mapped_lum / luminance, 1.0)
        
        for c in range(3):
            gpu_frame[:, :, c] *= ratio
        
        return cp.clip(gpu_frame, 0, 255)
    
    def set_mode(self, mode: str):
        """Set GameMateAI visual enhancement mode"""
        modes = {
            'fps': {'brightness': 1.3, 'saturation': 1.4, 'contrast': 1.3},
            'moba': {'brightness': 1.1, 'saturation': 1.2, 'contrast': 1.1},
            'rpg': {'brightness': 1.0, 'saturation': 1.3, 'contrast': 1.0}
        }
        
        if mode in modes:
            self.current_mode = modes[mode]
            logger.info(f"GameMateAI Visual mode set to: {mode}")
    
    def toggle(self):
        """Toggle GameMateAI visual enhancement"""
        self.enabled = not self.enabled
        logger.info(f"GameMateAI Visual Engine {'enabled' if self.enabled else 'disabled'}")